refactor(walk-log): 위젯 추출
