+++ lib/screens/walk_log_screen.dart
@@ -5,8 +5,11 @@ import 'package:wooyoungsoo/models/recently_walk_log_model.dart';
 import 'package:wooyoungsoo/services/walk_log_service/walk_log_service.dart';
 import 'package:wooyoungsoo/utils/constants.dart';
 import 'package:wooyoungsoo/widgets/common/navigation_bar_widget.dart';
+import 'package:wooyoungsoo/widgets/walk_log_screen/monthly_walking_log_view_widget.dart';
+import 'package:wooyoungsoo/widgets/walk_log_screen/walking_farm_color_bar_widget.dart';
+import 'package:wooyoungsoo/widgets/walk_log_screen/walking_farm_row_widget.dart';
 
-/// 마이페이지 화면
+/// 산책 로그 화면
 ///
 /// [screenWidth], [screenHeight] 화면의 너비와 높이
 /// [currentIndex] 현재 화면의 인덱스
@@ -17,6 +20,11 @@ class WalkLogScreen extends StatefulWidget {
   State<WalkLogScreen> createState() => _WalkLogScreenState();
 }
 
+/// 산책 로그 화면의 상태
+///
+/// [walkLogService] 산책 로그 서비스
+/// [recentlyWalkLogs] 최근 30일 산책 로그 리스트
+/// [monthlyWalkLogs] 이번 달 산책 로그 리스트
 class _WalkLogScreenState extends State<WalkLogScreen> {
   final WalkLogService walkLogService = WalkLogService();
   List<RecentlyWalkLogModel> recentlyWalkLogs = [];
@@ -29,12 +37,6 @@ class _WalkLogScreenState extends State<WalkLogScreen> {
       year: DateTime.now().year,
       month: DateTime.now().month,
     );
-
-    for (var i = 0; i < recentlyWalkLogs.length; i++) {
-      print(recentlyWalkLogs[i].walkDate);
-      print(recentlyWalkLogs[i].totalWalkedDistance);
-      print("=============");
-    }
 
     setState(() {});
   }
@@ -157,85 +159,7 @@ class _WalkLogScreenState extends State<WalkLogScreen> {
                         const SizedBox(
                           height: 20,
                         ),
-                        Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            const Text(
-                              "0km",
-                              style: TextStyle(
-                                fontSize: 12,
-                                fontWeight: fontWeightMedium,
-                                color: darkGreyColor,
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            Container(
-                              width: 20,
-                              height: 20,
-                              decoration: BoxDecoration(
-                                color: blueColorLv1,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            Container(
-                              width: 20,
-                              height: 20,
-                              decoration: BoxDecoration(
-                                color: blueColorLv2,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            Container(
-                              width: 20,
-                              height: 20,
-                              decoration: BoxDecoration(
-                                color: blueColorLv3,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            Container(
-                              width: 20,
-                              height: 20,
-                              decoration: BoxDecoration(
-                                color: blueColorLv4,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            Container(
-                              width: 20,
-                              height: 20,
-                              decoration: BoxDecoration(
-                                color: blueColorLv5,
-                                borderRadius: BorderRadius.circular(6),
-                              ),
-                            ),
-                            const SizedBox(
-                              width: 4,
-                            ),
-                            const Text(
-                              "3km",
-                              style: TextStyle(
-                                fontSize: 12,
-                                fontWeight: fontWeightMedium,
-                                color: darkGreyColor,
-                              ),
-                            ),
-                          ],
-                        ),
+                        const WalkingFarmColorBar(),
                       ],
                     ),
             ),
@@ -291,244 +215,13 @@ class _WalkLogScreenState extends State<WalkLogScreen> {
             const SizedBox(
               height: 20,
             ),
-            Padding(
-              padding: EdgeInsets.symmetric(
-                horizontal: screenWidth * 0.033,
-              ),
-              child: ListView.separated(
-                shrinkWrap: true,
-                physics: const NeverScrollableScrollPhysics(),
-                itemBuilder: (context, index) {
-                  return Container(
-                    decoration: BoxDecoration(
-                      color: whiteColor,
-                      borderRadius: BorderRadius.circular(20),
-                      boxShadow: const [
-                        BoxShadow(
-                          color: boxGreyColor,
-                          blurRadius: 50,
-                          offset: Offset(0, 0),
-                        ),
-                      ],
-                    ),
-                    child: Padding(
-                      padding: const EdgeInsets.all(10),
-                      child: Row(
-                        children: [
-                          monthlyWalkLogs[index].walkingImageUrl == null
-                              ? Container(
-                                  width: 100,
-                                  height: 100,
-                                  decoration: BoxDecoration(
-                                    borderRadius: BorderRadius.circular(20),
-                                    image: const DecorationImage(
-                                      image: AssetImage(
-                                          "assets/images/default_dog_image.png"),
-                                      fit: BoxFit.cover,
-                                    ),
-                                  ),
-                                )
-                              : Container(
-                                  width: 100,
-                                  height: 100,
-                                  decoration: BoxDecoration(
-                                    borderRadius: BorderRadius.circular(20),
-                                    image: DecorationImage(
-                                      image: NetworkImage(monthlyWalkLogs[index]
-                                          .walkingImageUrl!),
-                                      fit: BoxFit.cover,
-                                    ),
-                                  ),
-                                ),
-                          const SizedBox(
-                            width: 20,
-                          ),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Text(
-                                  monthlyWalkLogs[index]
-                                      .startedTime
-                                      .toString()
-                                      .split(" ")[0],
-                                  style: const TextStyle(
-                                    fontSize: 15,
-                                    fontWeight: fontWeightBold,
-                                    color: mainColor,
-                                  ),
-                                ),
-                                const SizedBox(
-                                  height: 5,
-                                ),
-                                const Divider(
-                                  color: lightGreyColor,
-                                  thickness: 1,
-                                ),
-                                const SizedBox(
-                                  height: 5,
-                                ),
-                                const Row(
-                                  children: [
-                                    SizedBox(
-                                      width: 80,
-                                      child: Text(
-                                        "산책 시간",
-                                        style: TextStyle(
-                                          fontSize: 12,
-                                          fontWeight: fontWeightMedium,
-                                          color: darkGreyColor,
-                                        ),
-                                      ),
-                                    ),
-                                    Text(
-                                      "거리 (km)",
-                                      style: TextStyle(
-                                        fontSize: 12,
-                                        fontWeight: fontWeightMedium,
-                                        color: darkGreyColor,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                                Row(
-                                  children: [
-                                    SizedBox(
-                                      width: 80,
-                                      child: Text(
-                                        monthlyWalkLogs[index]
-                                                    .walkingTime
-                                                    .toString()
-                                                    .split(".")[0][0] ==
-                                                "0"
-                                            ? "0${monthlyWalkLogs[index].walkingTime.toString().split(".")[0]}"
-                                            : monthlyWalkLogs[index]
-                                                .walkingTime
-                                                .toString()
-                                                .split(".")[0],
-                                        style: const TextStyle(
-                                          fontSize: 15,
-                                          fontWeight: fontWeightMedium,
-                                          color: blackColor,
-                                        ),
-                                      ),
-                                    ),
-                                    Text(
-                                      "${monthlyWalkLogs[index].walkedDistance}km",
-                                      style: const TextStyle(
-                                        fontSize: 15,
-                                        fontWeight: fontWeightMedium,
-                                        color: blackColor,
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ],
-                            ),
-                          ),
-                          const SizedBox(
-                            width: 10,
-                          ),
-                        ],
-                      ),
-                    ),
-                  );
-                },
-                separatorBuilder: (context, index) {
-                  return const Gap();
-                },
-                itemCount: monthlyWalkLogs.length,
-              ),
+            MonthlyWalkingLogView(
+              screenWidth: screenWidth,
+              monthlyWalkLogs: monthlyWalkLogs,
             ),
           ],
         ),
       ),
-    );
-  }
-}
-
-class WalkingFarmRow extends StatelessWidget {
-  const WalkingFarmRow({
-    super.key,
-    required this.walkLogs,
-  });
-
-  final List<RecentlyWalkLogModel> walkLogs;
-
-  Color getWalkLogColor(double walkedDistance) {
-    if (walkedDistance == 0) {
-      return blueColorLv1;
-    }
-    if (walkedDistance < 1) {
-      return blueColorLv2;
-    }
-    if (walkedDistance < 2) {
-      return blueColorLv3;
-    }
-    if (walkedDistance < 3) {
-      return blueColorLv4;
-    }
-    return blueColorLv5;
-  }
-
-  Color getDayColor(double walkedDistance) {
-    if (walkedDistance < 2) {
-      return darkGreyColor;
-    }
-    return whiteColor;
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return SizedBox(
-      width: 270,
-      height: 30,
-      child: ListView.separated(
-        scrollDirection: Axis.horizontal,
-        itemBuilder: (context, index) {
-          return Container(
-            width: 30,
-            height: 30,
-            decoration: BoxDecoration(
-              color: getWalkLogColor(walkLogs[index].totalWalkedDistance),
-              borderRadius: BorderRadius.circular(6),
-            ),
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                Text(
-                  walkLogs[index].walkDate.day.toString(),
-                  textAlign: TextAlign.center,
-                  style: TextStyle(
-                    fontSize: 13,
-                    fontWeight: fontWeightBold,
-                    color: getDayColor(walkLogs[index].totalWalkedDistance),
-                  ),
-                ),
-              ],
-            ),
-          );
-        },
-        separatorBuilder: (context, index) {
-          return const SizedBox(
-            width: 10,
-          );
-        },
-        itemCount: walkLogs.length,
-      ),
-    );
-  }
-}
-
-class Gap extends StatelessWidget {
-  const Gap({
-    super.key,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return const SizedBox(
-      height: 10,
     );
   }
 }
